change: rework and keep only the test, move other functions to fixtures or utils
